requirements.txt
django==2.0
pyyaml==3.12
pillow==4.3.0
django_compressor==2.2
django-libsass==0.7
rules==1.3
django-tellme==0.6.5
# for messaging
coverage==3.7.1
django-coverage==1.2.4
django-widget-tweaks==1.4.2
django-watson==1.5.2 #Search engine
django-cleanup==2.1.0 #To automatically delete old files when new ones are uploaded (profile picture for example)
social-auth-app-django==2.1.0